+++ ta4j-core/src/main/java/org/ta4j/core/BaseBar.java
@@ -49,9 +49,9 @@ public class BaseBar implements Bar {
     /** Close price of the period */
     private Num closePrice = null;
     /** Max price of the period */
-    private Num maxPrice = null;
+    private Num highPrice = null;
     /** Min price of the period */
-    private Num minPrice = null;
+    private Num lowPrice = null;
     /** Traded amount during the period */
     private Num amount;
     /** Volume of the period */
@@ -110,6 +110,25 @@ public class BaseBar implements Bar {
 
     /**
      * Constructor.
+     * @param endTime the end time of the bar
+     * @param openPrice the open price of the bar
+     * @param highPrice the highest price of the bar
+     * @param lowPrice the lowest price of the bar
+     * @param closePrice the close price of the bar
+     * @param volume the volume of the bar
+     * @param value the value of the bar
+     */
+    public BaseBar(ZonedDateTime endTime, String openPrice, String highPrice, String lowPrice, String closePrice, String volume, String value, Function<Number, Num> numFunction) {
+        this(endTime, numFunction.apply(new BigDecimal(openPrice)),
+                numFunction.apply(new BigDecimal(highPrice)),
+                numFunction.apply(new BigDecimal(lowPrice)),
+                numFunction.apply(new BigDecimal(closePrice)),
+                numFunction.apply(new BigDecimal(volume)),
+                numFunction.apply(new BigDecimal(value)));
+    }
+
+    /**
+     * Constructor.
      * @param endTime the end time of the bar period
      * @param openPrice the open price of the bar period
      * @param highPrice the highest price of the bar period
@@ -139,8 +158,8 @@ public class BaseBar implements Bar {
         this.endTime = endTime;
         this.beginTime = endTime.minus(timePeriod);
         this.openPrice = openPrice;
-        this.maxPrice = highPrice;
-        this.minPrice = lowPrice;
+        this.highPrice = highPrice;
+        this.lowPrice = lowPrice;
         this.closePrice = closePrice;
         this.volume = volume;
         this.amount = amount;
@@ -157,14 +176,14 @@ public class BaseBar implements Bar {
      * @return the low price of the period
      */
     public Num getLowPrice() {
-        return minPrice;
+        return lowPrice;
     }
 
     /**
      * @return the high price of the period
      */
     public Num getHighPrice() {
-        return maxPrice;
+        return highPrice;
     }
 
     /**
@@ -236,22 +255,22 @@ public class BaseBar implements Bar {
         }
 
         closePrice = price;
-        if (maxPrice == null) {
-            maxPrice = price;
-        } else if(maxPrice.isLessThan(price)) {
-            maxPrice = price;
+        if (highPrice == null) {
+            highPrice = price;
+        } else if(highPrice.isLessThan(price)) {
+            highPrice = price;
         }
-        if (minPrice == null) {
-            minPrice = price;
-        } else if(minPrice.isGreaterThan(price)){
-            minPrice = price;
+        if (lowPrice == null) {
+            lowPrice = price;
+        } else if(lowPrice.isGreaterThan(price)){
+            lowPrice = price;
         }
     }
 
     @Override
     public String toString() {
         return String.format("{end time: %1s, close price: %2$f, open price: %3$f, min price: %4$f, max price: %5$f, volume: %6$f}",
-                endTime.withZoneSameInstant(ZoneId.systemDefault()), closePrice.doubleValue(), openPrice.doubleValue(), minPrice.doubleValue(), maxPrice.doubleValue(), volume.doubleValue());
+                endTime.withZoneSameInstant(ZoneId.systemDefault()), closePrice.doubleValue(), openPrice.doubleValue(), lowPrice.doubleValue(), highPrice.doubleValue(), volume.doubleValue());
     }
 
     /**
@@ -266,5 +285,103 @@ public class BaseBar implements Bar {
         if (endTime == null) {
             throw new IllegalArgumentException("End time cannot be null");
         }
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null) {
+            return false;
+        }
+        if (!(obj instanceof BaseBar)) {
+            return false;
+        }
+        BaseBar other = (BaseBar) obj;
+        if (amount == null) {
+            if (other.amount != null) {
+                return false;
+            }
+        } else if (!amount.equals(other.amount)) {
+            return false;
+        }
+        if (beginTime == null) {
+            if (other.beginTime != null) {
+                return false;
+            }
+        } else if (!beginTime.equals(other.beginTime)) {
+            return false;
+        }
+        if (closePrice == null) {
+            if (other.closePrice != null) {
+                return false;
+            }
+        } else if (!closePrice.equals(other.closePrice)) {
+            return false;
+        }
+        if (endTime == null) {
+            if (other.endTime != null) {
+                return false;
+            }
+        } else if (!endTime.equals(other.endTime)) {
+            return false;
+        }
+        if (highPrice == null) {
+            if (other.highPrice != null) {
+                return false;
+            }
+        } else if (!highPrice.equals(other.highPrice)) {
+            return false;
+        }
+        if (lowPrice == null) {
+            if (other.lowPrice != null) {
+                return false;
+            }
+        } else if (!lowPrice.equals(other.lowPrice)) {
+            return false;
+        }
+        if (openPrice == null) {
+            if (other.openPrice != null) {
+                return false;
+            }
+        } else if (!openPrice.equals(other.openPrice)) {
+            return false;
+        }
+        if (timePeriod == null) {
+            if (other.timePeriod != null) {
+                return false;
+            }
+        } else if (!timePeriod.equals(other.timePeriod)) {
+            return false;
+        }
+        if (trades != other.trades) {
+            return false;
+        }
+        if (volume == null) {
+            if (other.volume != null) {
+                return false;
+            }
+        } else if (!volume.equals(other.volume)) {
+            return false;
+        }
+        return true;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((amount == null) ? 0 : amount.hashCode());
+        result = prime * result + ((beginTime == null) ? 0 : beginTime.hashCode());
+        result = prime * result + ((closePrice == null) ? 0 : closePrice.hashCode());
+        result = prime * result + ((endTime == null) ? 0 : endTime.hashCode());
+        result = prime * result + ((highPrice == null) ? 0 : highPrice.hashCode());
+        result = prime * result + ((lowPrice == null) ? 0 : lowPrice.hashCode());
+        result = prime * result + ((openPrice == null) ? 0 : openPrice.hashCode());
+        result = prime * result + ((timePeriod == null) ? 0 : timePeriod.hashCode());
+        result = prime * result + trades;
+        result = prime * result + ((volume == null) ? 0 : volume.hashCode());
+        return result;
     }
 }