+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/LowPriceIndicator.java
@@ -23,22 +23,17 @@
  *******************************************************************************/
 package org.ta4j.core.indicators.helpers;
 
+import org.ta4j.core.Bar;
 import org.ta4j.core.TimeSeries;
-import org.ta4j.core.indicators.CachedIndicator;
-import org.ta4j.core.num.Num;
 
 /**
  * Low price indicator.
  * </p>
  */
-public class LowPriceIndicator extends CachedIndicator<Num> {
+public class LowPriceIndicator extends PriceIndicator {
 
     public LowPriceIndicator(TimeSeries series) {
-        super(series);
+        super(series, Bar::getLowPrice);
     }
 
-    @Override
-    protected Num calculate(int index) {
-        return getTimeSeries().getBar(index).getLowPrice();
-    }
 }